Moved new discovery classes and tests to discovery package
