+++ junit5-engine/src/main/java/org/junit/gen5/engine/junit5/discovery/DiscoverySelectorResolver.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.junit.gen5.engine.junit5.discoveryNEW;
+package org.junit.gen5.engine.junit5.discovery;
 
 import static java.lang.String.format;
 import static org.junit.gen5.commons.util.ReflectionUtils.*;
@@ -25,7 +25,6 @@ import org.junit.gen5.engine.discovery.ClasspathSelector;
 import org.junit.gen5.engine.discovery.MethodSelector;
 import org.junit.gen5.engine.discovery.PackageSelector;
 import org.junit.gen5.engine.discovery.UniqueIdSelector;
-import org.junit.gen5.engine.junit5.discovery.IsScannableTestClass;
 
 public class DiscoverySelectorResolver {
