+++ junit5-engine/src/main/java/org/junit/gen5/engine/junit5/discovery/ElementResolver.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.junit.gen5.engine.junit5.discoveryNEW;
+package org.junit.gen5.engine.junit5.discovery;
 
 import java.lang.reflect.AnnotatedElement;
 import java.util.Optional;
