--- junit5-engine/src/main/java/org/junit/gen5/engine/junit5/discoveryNEW/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Internal classes for test discovery within the JUnit 5 test engine.
- */
-
-package org.junit.gen5.engine.junit5.discoveryNEW;
